+++ package.json
@@ -7,11 +7,27 @@
   "type": "module",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
+  "sideEffects": false,
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js",
       "default": "./dist/index.js"
+    },
+    "./browser": {
+      "types": "./dist/index.browser.d.ts",
+      "import": "./dist/index.browser.js",
+      "default": "./dist/index.browser.js"
+    },
+    "./network/http": {
+      "types": "./dist/network/http.d.ts",
+      "import": "./dist/network/http.js",
+      "default": "./dist/network/http.js"
+    },
+    "./network/browser": {
+      "types": "./dist/network/browser.d.ts",
+      "import": "./dist/network/browser.js",
+      "default": "./dist/network/browser.js"
     },
     "./package.json": "./package.json"
   },
