+++ src/resonate.ts
@@ -1,5 +1,3 @@
-import { randomUUID } from "node:crypto";
-import { setTimeout as delay } from "node:timers/promises";
 import { WallClock } from "./clock.js";
 import { Codec } from "./codec.js";
 import { Core } from "./core.js";
@@ -7,7 +5,7 @@ import { type Encryptor, NoopEncryptor } from "./encryptor.js";
 import exceptions, { ResonateTimeoutException } from "./exceptions.js";
 import { AsyncHeartbeat, type Heartbeat, NoopHeartbeat } from "./heartbeat.js";
 import { ConsoleLogger, type Logger, type LogLevel } from "./logger.js";
-import { HttpNetwork, PollMessageSource } from "./network/http.js";
+import { HttpNetwork, PollMessageSource } from "./network/browser.js";
 import { LocalNetwork } from "./network/local.js";
 import type { Network } from "./network/network.js";
 import {
@@ -22,6 +20,7 @@ import {
   type TaskRecord,
 } from "./network/types.js";
 import { type Options, OptionsBuilder } from "./options.js";
+import { delay, getEnv, randomUUID } from "./platform.js";
 import { Promises } from "./promises.js";
 import { Registry } from "./registry.js";
 import { Schedules } from "./schedules.js";
@@ -129,7 +128,7 @@ export class Resonate {
     this.ttl = ttl;
     this.codec = new Codec(encryptor ?? new NoopEncryptor());
 
-    const resolvedPrefix = prefix ?? process.env.RESONATE_PREFIX;
+    const resolvedPrefix = prefix ?? getEnv("RESONATE_PREFIX");
     this.idPrefix = resolvedPrefix ? `${resolvedPrefix}:` : "";
 
     // Resolve logger: explicit logger > ConsoleLogger with resolved level
@@ -144,9 +143,9 @@ export class Resonate {
     // Full URL/bun fmt
     // bun check
     // /token/timeout env var resolution is HttpNetwork's responsibility (3.7).
-    const resolvedUrl = url ?? (process.env.RESONATE_URL || undefined);
+    const resolvedUrl = url ?? (getEnv("RESONATE_URL") || undefined);
 
-    this.pid = pid ?? crypto.randomUUID().replace(/-/g, "");
+    this.pid = pid ?? randomUUID().replace(/-/g, "");
 
     let heartbeat: boolean;
     if (network) {
