+++ src/util.ts
@@ -1,4 +1,3 @@
-import { randomUUID } from "node:crypto";
 import type { Codec } from "./codec.js";
 import type { InnerContext } from "./context.js";
 import exceptions from "./exceptions.js";
@@ -12,6 +11,7 @@ import {
   type PromiseSettleRes,
 } from "./network/types.js";
 import { type Options, RESONATE_OPTIONS } from "./options.js";
+import { randomUUID } from "./platform.js";
 import type { Effects, Func, Result, Send } from "./types.js";
 
 // version
